Refactored to actually have a name instead of test_gui
